Solved time_between_pulses issue
there used to be a large delay between pulses, even when the time was set to zero. this was caused by a +2 i had placed next to the while loop. the solution was to lower this value to 0.001, which allows the code to run. if you add no additional time, the loop ends before the pins are turned off. this is the smallest time i tested, you may be able to shorten this even more if needed.